Move and fix kafka tests
